Added tree biaser
This is for correcting errors such as "(1 - (2 - 3))" while the correct
answer should be "((1 - 2) - 3)". This is applicable to exponentiation and
other forms as well.

Currently only one type of bias is included: BiasLeft. It assumes that the
production involves three RHS items, the first being the terminal operator
node.

TODO:
1. In GraphicalProduction consturctors, check to make sure that only
appropriate productions can be labeled as Biasleft.

2. Add LHS member variable to GraphicalProduction, so that we don't need
to use String split every time. This will server to improve performance.