Added new parameter recursionGeomScoreRatioThresh to tokenSetParser
This value, when set to a value > 0.0 and < 1.0, can speed up the parsing
process appreciably, by a factor of 2 - 8, depending on whether it is
tested on a PC or on the Android phone.

It seems that using a value as large as 0.90 does not cause problems, at
least under the current relatively small grammar set. But it needs to be
watched closely.