Adding support for brackets.
Currently, brackets can be used on ADDITION and MULTIPLICATION only.
Further expansion of support for brackets may be required. This can be
done by adding new productions to "productions.txt" or by writing code
that automatically append bracket productions given a bracket-free grammar
file.

TODO: Brackets tend to be merged with near by tokens in StrokeCurator.
Perhaps we need to write a set of heuristic rules for preventing the
merging in StrokeCurator.