NodeToken and incremental parsing
AbstractToken is now the parent class of CWrittenToken and NodeToken.
NodeToken is a newly added class that contains a written token set and the
parsed node that corresponds to this token set. It can be a part of any
CWrittenTokenSetNoStrokes and co-exist with other AbstractTokens (either
CWrittenToken or NodeToken). The parser was augmented to handle both types
of tokens and there mixes.

Unit-test cases were added.

Note: The evaluator part of NodeToken parsing still has bugs. For example
if "A" is already parsed as a NodeToken and included in a token set like
"A = 3", the evaluator will fail due to the evaluation of "A" into a
value.