Partially fixing the issue with requiredTermTypes
The cyclic nature of the production set makes it hard to determined the required terminal types of each production. This change set improves the result of the rrequired-term-type calculation a little through examination of the call stack and tweaking in productions.txt to make sure that there is no "advance" use of productions, i.e., use of a production as a child of another production before this production is defined.

The newly added test Test_GraphicalProductionSet ties down some of the most important requird-term-type expectations given the current production set.

Due to this improvement, the parser speed is enhanced by about 20%.